+++ CHANGELOG.md
@@ -5,6 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 3.0.0 - 2025-04-07
+
+### Added
+
+- Support for the v2 autocomplete and place details APIs!
+- **BREAKING:** We have renamed the Place Details method to clarify its purpose.
+
+If you want to keep using the v1 endpoint, you can amend your code like so:
+
+```diff
+- $result = $this->apiInstance->place([$this->kultuurikatel['gid']]);
++ $result = $this->apiInstance->placeDetails([$this->kultuurikatel['gid']]);
+```
+
+To upgrade to the v2 Place Details endpoint, which features improved address formatting,
+use the new V2 method:
+
+```diff
+- $result = $this->apiInstance->place([$this->kultuurikatel['gid']]);
++ $result = $this->apiInstance->placeDetailsV2([$this->kultuurikatel['gid']]);
+```
+
+For an overview of the structural changes we've made in the V2 API,
+refer to the [migration guide](https://docs.stadiamaps.com/geocoding-search-autocomplete/v2-api-migration-guide/).
+
 ## Version 2.0.0 - 2025-01-26
 
 ### Added